fix: immediate character data loading and derived attributes float support
- Remove tick_update wait in character.js - data via HTTP API is immediately available
- Add derived_attributes: HashMap<String, f32> to AgentSelfState protocol
- Change evaluate_max_value return type i32 -> f32
- Change DerivedAttributeComponent cache/calculate to use f32
- Update broadcaster to fill derived_attributes field
- Update frontend to display float values with 2 decimal places
- Add derived_raw field to AttributesGlimpse for programmatic access

Breaks: AgentSelfState protocol (new field added)

+++ crates/server/src/handlers/dashboard.rs
@@ -586,9 +586,9 @@ pub async fn get_agent_details(
             {
                 return crate::game_data::types::StatusComponent::evaluate_max_value(
                     &attr_def.max_value_formula,
-                    100,
+                    100.0,
                     &attributes_map,
-                );
+                ) as i32;
             }
             100
         };
